[*] Build: Improve default environment for using it in build scripts, includeing internal (src/elecard/firmware/bin)

+++ src/script/006_move_devVersion_file.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-mv -f $PRJROOT/src/update/.development_revision $PRJROOT/src/firmware/.development_revision
+if [ -e $PRJROOT/src/update/.development_revision ]; then
+	mv -f $PRJROOT/src/update/.development_revision $PRJROOT/src/firmware/.development_revision
+fi
